+++ features/admin.applications.v1/components/settings/shared-access.tsx
@@ -16,27 +16,15 @@
  * under the License.
  */
 
-import Alert from "@oxygen-ui/react/Alert";
-import AlertTitle from "@oxygen-ui/react/AlertTitle";
-import Box from "@oxygen-ui/react/Box";
-import Button from "@oxygen-ui/react/Button";
 import { FeatureConfigInterface } from "@wso2is/admin.core.v1/models/config";
 import { IdentifiableComponentInterface, SBACInterface } from "@wso2is/core/models";
-import { ConfirmationModal, EmphasizedSegment } from "@wso2is/react-components";
+import { EmphasizedSegment } from "@wso2is/react-components";
 import React, { FunctionComponent, ReactElement, useState } from "react";
-import { Trans, useTranslation } from "react-i18next";
 import { useOperationStatusPoller } from "../../hooks/use-operation-status-poller";
 import { ApplicationInterface } from "../../models/application";
 import { ApplicationShareForm } from "../forms/share-application-form";
-import ApplicationShareStatusWizard from "../wizard/application-share-status-wizard";
-import { ModalWithSidePanel } from "@wso2is/admin.core.v1/components/modals/modal-with-side-panel";
-import {
-    Heading,
-    LinkButton,
-} from "@wso2is/react-components";
-import { Grid } from "semantic-ui-react";
-
-export type OperationStatus = "IDLE" | "ONGOING" | "SUCCESS" | "FAILED" | "PARTIAL";
+import { OperationStatusBanner } from "../shared-access-status-banner";
+import { ApplicationShareStatus } from "../../constants/application-management";
 
 /**
  * Proptypes for the shared access component.
@@ -67,22 +55,15 @@ export const SharedAccess: FunctionComponent<SharedAccessPropsInterface> = (
     props: SharedAccessPropsInterface
 ): ReactElement => {
 
-    const { t } = useTranslation();
-
     const { application, onUpdate, readOnly } = props;
-    const [ sharingState, setSharingState ] = useState<OperationStatus>("IDLE");
+    const [ sharingState, setSharingState ] = useState<ApplicationShareStatus>(ApplicationShareStatus.IDLE);
     const [ sharingOperationId, setSharingOperationId ] = useState<string>();
-    const [ showBanner, setShowBanner ] = useState(false);
-    const [ showConfirmationModal, setShowConfirmationModal ] = useState(false);
-    const [ showStatusModal, setShowStatusModal ] = useState(false);
-    const [ componentId, SetComponentId ] = useState();
 
     const { status, startPolling } = useOperationStatusPoller({  // use-application-sharing-operation-status
         applicationId: application.id,
         pollingInterval: 5000,
         onCompleted: (finalStatus) => {
             setSharingState(finalStatus);
-            setShowBanner(true);
         },
         onStatusChange: (newOperationId, newStatus) => {
             setSharingState(newStatus);
@@ -90,142 +71,20 @@ export const SharedAccess: FunctionComponent<SharedAccessPropsInterface> = (
         },
         onError: (error) => {
             console.error("Polling error:", error);
-        }//enabled
+        },
+        enabled: true //config.operationStatusPoller?.enabled ??
     });
 
     const handleChildStartedOperation = () => {
-        setSharingState("ONGOING");
-        setShowBanner(true);
+        setSharingState(ApplicationShareStatus.IN_PROGRESS);
         startPolling();
     };
-    
-    const handleShowStatusModal = () => {
-        setShowStatusModal(true);
-    };
-    const handleHideStatusModal = () => {
-        setShowStatusModal(false);
-    };
-
-    const resolveContent = (): ReactElement => {
-        return (
-            <ApplicationShareStatusWizard
-                componentId="wizard1"
-                operationId={sharingOperationId}
-                isSubmitting={true}
-                isLoading={true}
-                hasError={false}
-            />
-        );
-    };
-
-    const renderBanner = (): ReactElement | null => {
-
-        const isOngoing = sharingState === "ONGOING";
-        const isFailure = sharingState === "PARTIAL" || sharingState === "FAILED";
-
-        if ( !isOngoing && !isFailure ) return null;
-        return (
-            <>
-                { sharingState === "ONGOING" && (
-                    <div className="banner-wrapper">
-                        <Alert
-                            severity="warning"
-                        >
-                            <AlertTitle className="alert-title">
-                                <Trans components={ { strong: <strong/> } } >
-                                    Update In Progress.
-                                </Trans>
-                            </AlertTitle>
-                            Updating shared access is in progress.
-                        </Alert>
-                    </div>
-                ) }
-                { isFailure && (
-                    <div className="banner-wrapper">
-                        <Alert
-                            // className={ classes }
-                            severity="warning"
-                            action={
-                                (
-                                    <Box display="flex">
-                                        <Button
-                                            className="banner-view-hide-details"
-                                            onClick={handleShowStatusModal}>
-                                            {
-                                                "View"
-                                            }
-                                        </Button>
-                                    </Box>
-                                )
-                            }
-                        >
-                            <AlertTitle className="alert-title">
-                                <Trans components={ { strong: <strong/> } } >
-                                    Update Partialy Successfull.
-                                </Trans>
-                            </AlertTitle>
-                            Updating shared access completed with partial success.
-                        </Alert>
-                    </div>
-                ) }
-            </>
-            
-        );
-    };
-
-    
 
     return (
         <>
-            { renderBanner() }
-
-            <ModalWithSidePanel
-                data-testid={ componentId }
-                data-componentid={ componentId }
-                open={ showStatusModal }
-                className="wizard application-create-wizard"
-                dimmer="blurring"
-                size="small"
-                onClose={ () => handleHideStatusModal() }
-                closeOnDimmerClick={ false }
-                closeOnEscape
-            >
-                <ModalWithSidePanel.MainPanel>
-                    <ModalWithSidePanel.Header className="wizard-header">
-                            Summary - Update application shared access
-                        <Heading as="h6">
-                            Summary of detailed application sharing failures.
-                        </Heading>
-                    </ModalWithSidePanel.Header>
-    
-                    <ModalWithSidePanel.Content className="content-container">
-                        <Grid>
-                            { resolveContent() }
-                        </Grid>
-    
-                    </ModalWithSidePanel.Content>
-                    <ModalWithSidePanel.Actions>
-                        <Grid>
-                            <Grid.Row column={ 1 }>
-                                <Grid.Column mobile={ 8 } tablet={ 8 } computer={ 8 }>
-                                    <LinkButton
-                                        data-testid={ `-close-button` }
-                                        data-componentid={ `-close-button` }
-                                        floated="left"
-                                        onClick={ () => {
-                                            handleHideStatusModal();
-                                        } }
-                                        disabled={ false }
-                                    >
-                                        close
-                                    </LinkButton>
-                                </Grid.Column>
-                            </Grid.Row>                           
-                        </Grid>
-                    </ModalWithSidePanel.Actions>                
-                </ModalWithSidePanel.MainPanel>
-            </ModalWithSidePanel>
-            
+            <OperationStatusBanner
+                status={ sharingState}
+                sharingOperationId={ sharingOperationId }/>
             <EmphasizedSegment className="advanced-configuration-section" padded="very">
                 <ApplicationShareForm
                     application={ application }
@@ -233,7 +92,7 @@ export const SharedAccess: FunctionComponent<SharedAccessPropsInterface> = (
                     readOnly={ readOnly }
                     onOperationStarted={ handleChildStartedOperation }
                     operationStatus={ status }
-                    isSharingInProgress={ sharingState === "ONGOING" }
+                    isSharingInProgress={ sharingState === ApplicationShareStatus.IN_PROGRESS }
                 />
             </EmphasizedSegment>
         </>
